% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proximitybeacon_objects.R
\name{AdvertisedId}
\alias{AdvertisedId}
\title{AdvertisedId Object}
\usage{
AdvertisedId(type = NULL, id = NULL)
}
\arguments{
\item{type}{Specifies the identifier type}

\item{id}{The actual beacon identifier, as broadcast by the beacon hardware}
}
\value{
AdvertisedId object
}
\description{
AdvertisedId Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Defines a unique identifier of a beacon as broadcast by the device.
}
