% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proximitybeacon_objects.R
\name{AttachmentInfo}
\alias{AttachmentInfo}
\title{AttachmentInfo Object}
\usage{
AttachmentInfo(namespacedType = NULL, data = NULL)
}
\arguments{
\item{namespacedType}{Specifies what kind of attachment this is}

\item{data}{An opaque data container for client-provided data}
}
\value{
AttachmentInfo object
}
\description{
AttachmentInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A subset of attachment information served via the `beaconinfo.getforobserved` method, used when your users encounter your beacons.
}

