% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proximitybeacon_objects.R
\name{Beacon.properties}
\alias{Beacon.properties}
\title{Beacon.properties Object}
\usage{
Beacon.properties()
}
\value{
Beacon.properties object
}
\description{
Beacon.properties Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Properties of the beacon device, for example battery type or firmware version. Optional.
}
\seealso{
Other Beacon functions: \code{\link{Beacon}},
  \code{\link{beacons.register}},
  \code{\link{beacons.update}}
}
