% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proximitybeacon_objects.R
\name{Beacon}
\alias{Beacon}
\title{Google Proximity Beacon API Objects 
Registers, manages, indexes, and searches beacons.}
\usage{
Beacon(Beacon.properties = NULL, beaconName = NULL, advertisedId = NULL,
  status = NULL, placeId = NULL, latLng = NULL, indoorLevel = NULL,
  expectedStability = NULL, description = NULL, properties = NULL,
  ephemeralIdRegistration = NULL, provisioningKey = NULL)
}
\arguments{
\item{Beacon.properties}{The \link{Beacon.properties} object or list of objects}

\item{beaconName}{Resource name of this beacon}

\item{advertisedId}{The identifier of a beacon as advertised by it}

\item{status}{Current status of the beacon}

\item{placeId}{The [Google Places API](/places/place-id) Place ID of the place where the beacon is deployed}

\item{latLng}{The location of the beacon, expressed as a latitude and longitude pair}

\item{indoorLevel}{The indoor level information for this beacon, if known}

\item{expectedStability}{Expected location stability}

\item{description}{Free text used to identify and describe the beacon}

\item{properties}{Properties of the beacon device, for example battery type or firmware version}

\item{ephemeralIdRegistration}{Write-only registration parameters for beacons using Eddystone-EID (remotely resolved ephemeral ID) format}

\item{provisioningKey}{Some beacons may require a user to provide an authorization key before changing any of its configuration (e}
}
\value{
Beacon object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2017-03-05 19:58:07
filename: /Users/mark/dev/R/autoGoogleAPI/googleproximitybeaconv1beta1.auto/R/proximitybeacon_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
Beacon Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Details of a beacon device.
}
\seealso{
Other Beacon functions: \code{\link{Beacon.properties}},
  \code{\link{beacons.register}},
  \code{\link{beacons.update}}
}
