% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proximitybeacon_objects.R
\name{BeaconAttachment}
\alias{BeaconAttachment}
\title{BeaconAttachment Object}
\usage{
BeaconAttachment(attachmentName = NULL, namespacedType = NULL,
  data = NULL)
}
\arguments{
\item{attachmentName}{Resource name of this attachment}

\item{namespacedType}{Specifies what kind of attachment this is}

\item{data}{An opaque data container for client-provided data}
}
\value{
BeaconAttachment object
}
\description{
BeaconAttachment Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Project-specific data associated with a beacon.
}
