% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proximitybeacon_functions.R
\name{beaconinfo.getforobserved}
\alias{beaconinfo.getforobserved}
\title{Given one or more beacon observations, returns any beacon information and attachments accessible to your application. Authorize by using the [API key](https://developers.google.com/beacons/proximity/how-tos/authorizing#APIKey) for the application.}
\usage{
beaconinfo.getforobserved(GetInfoForObservedBeaconsRequest)
}
\arguments{
\item{GetInfoForObservedBeaconsRequest}{The \link{GetInfoForObservedBeaconsRequest} object to pass to this method
#' @importFrom googleAuthR gar_api_generator}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item 
}

Set \code{options(googleAuthR.scopes.selected = c()}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/beacons/proximity/}{Google Documentation}

Other GetInfoForObservedBeaconsRequest functions: \code{\link{GetInfoForObservedBeaconsRequest}}
}
