% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proximitybeacon_objects.R
\name{BeaconInfo}
\alias{BeaconInfo}
\title{BeaconInfo Object}
\usage{
BeaconInfo(advertisedId = NULL, beaconName = NULL, attachments = NULL)
}
\arguments{
\item{advertisedId}{The ID advertised by the beacon}

\item{beaconName}{The name under which the beacon is registered}

\item{attachments}{Attachments matching the type(s) requested}
}
\value{
BeaconInfo object
}
\description{
BeaconInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A subset of beacon information served via the `beaconinfo.getforobserved` method, which you call when users of your app encounter your beacons.
}
