% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proximitybeacon_functions.R
\name{beacons.list}
\alias{beacons.list}
\title{Searches the beacon registry for beacons that match the given search criteria. Only those beacons that the client has permission to list will be returned. Authenticate using an [OAuth access token](https://developers.google.com/identity/protocols/OAuth2) from a signed-in user with **viewer**, **Is owner** or **Can edit** permissions in the Google Developers Console project.}
\usage{
beacons.list(q = NULL, pageToken = NULL, pageSize = NULL,
  projectId = NULL)
}
\arguments{
\item{q}{Filter query string that supports the following field filters: * `description:''` For example: `description:'Room 3'` Returns beacons whose description matches tokens in the string 'Room 3' (not necessarily that exact string)}

\item{pageToken}{A pagination token obtained from a previous request to list beacons}

\item{pageSize}{The maximum number of records to return for this request, up to a server-defined upper limit}

\item{projectId}{The project id to list beacons under}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/userlocation.beacon.registry
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/userlocation.beacon.registry)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/beacons/proximity/}{Google Documentation}
}
