% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proximitybeacon_functions.R
\name{beacons.register}
\alias{beacons.register}
\title{Registers a previously unregistered beacon given its `advertisedId`. These IDs are unique within the system. An ID can be registered only once. Authenticate using an [OAuth access token](https://developers.google.com/identity/protocols/OAuth2) from a signed-in user with **Is owner** or **Can edit** permissions in the Google Developers Console project.}
\usage{
beacons.register(Beacon, projectId = NULL)
}
\arguments{
\item{Beacon}{The \link{Beacon} object to pass to this method}

\item{projectId}{The project id of the project the beacon will be registered to}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/userlocation.beacon.registry
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/userlocation.beacon.registry)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/beacons/proximity/}{Google Documentation}

Other Beacon functions: \code{\link{Beacon.properties}},
  \code{\link{Beacon}}, \code{\link{beacons.update}}
}
