% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proximitybeacon_functions.R
\name{beacons.update}
\alias{beacons.update}
\title{Updates the information about the specified beacon. **Any field that you do not populate in the submitted beacon will be permanently erased**, so you should follow the 'read, modify, write' pattern to avoid inadvertently destroying data. Changes to the beacon status via this method will be silently ignored. To update beacon status, use the separate methods on this API for activation, deactivation, and decommissioning. Authenticate using an [OAuth access token](https://developers.google.com/identity/protocols/OAuth2) from a signed-in user with **Is owner** or **Can edit** permissions in the Google Developers Console project.}
\usage{
beacons.update(Beacon, beaconName, projectId = NULL)
}
\arguments{
\item{Beacon}{The \link{Beacon} object to pass to this method}

\item{beaconName}{Resource name of this beacon}

\item{projectId}{The project id of the beacon to update}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/userlocation.beacon.registry
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/userlocation.beacon.registry)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/beacons/proximity/}{Google Documentation}

Other Beacon functions: \code{\link{Beacon.properties}},
  \code{\link{Beacon}}, \code{\link{beacons.register}}
}

