% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proximitybeacon_objects.R
\name{Date}
\alias{Date}
\title{Date Object}
\usage{
Date(year = NULL, month = NULL, day = NULL)
}
\arguments{
\item{year}{Year of date}

\item{month}{Month of year}

\item{day}{Day of month}
}
\value{
Date object
}
\description{
Date Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a whole calendar date, e.g. date of birth. The time of day and time zone are either specified elsewhere or are not significant. The date is relative to the Proleptic Gregorian Calendar. The day may be 0 to represent a year and month where the day is not significant, e.g. credit card expiration date. The year may be 0 to represent a month and day independent of year, e.g. anniversary date. Related types are google.type.TimeOfDay and `google.protobuf.Timestamp`.
}
