% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proximitybeacon_objects.R
\name{DeleteAttachmentsResponse}
\alias{DeleteAttachmentsResponse}
\title{DeleteAttachmentsResponse Object}
\usage{
DeleteAttachmentsResponse(numDeleted = NULL)
}
\arguments{
\item{numDeleted}{The number of attachments that were deleted}
}
\value{
DeleteAttachmentsResponse object
}
\description{
DeleteAttachmentsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response for a request to delete attachments.
}

