% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proximitybeacon_objects.R
\name{Diagnostics}
\alias{Diagnostics}
\title{Diagnostics Object}
\usage{
Diagnostics(beaconName = NULL, estimatedLowBatteryDate = NULL,
  alerts = NULL)
}
\arguments{
\item{beaconName}{Resource name of the beacon}

\item{estimatedLowBatteryDate}{The date when the battery is expected to be low}

\item{alerts}{An unordered list of Alerts that the beacon has}
}
\value{
Diagnostics object
}
\description{
Diagnostics Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Diagnostics for a single beacon.
}

