% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proximitybeacon_objects.R
\name{EphemeralIdRegistrationParams}
\alias{EphemeralIdRegistrationParams}
\title{EphemeralIdRegistrationParams Object}
\usage{
EphemeralIdRegistrationParams(serviceEcdhPublicKey = NULL,
  minRotationPeriodExponent = NULL, maxRotationPeriodExponent = NULL)
}
\arguments{
\item{serviceEcdhPublicKey}{The beacon service's public key for use by a beacon to derive its Identity Key using Elliptic Curve Diffie-Hellman key exchange}

\item{minRotationPeriodExponent}{Indicates the minimum rotation period supported by the service}

\item{maxRotationPeriodExponent}{Indicates the maximum rotation period supported by the service}
}
\value{
EphemeralIdRegistrationParams object
}
\description{
EphemeralIdRegistrationParams Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Information a client needs to provision and register beacons that broadcast Eddystone-EID format beacon IDs, using Elliptic curve Diffie-Hellman key exchange. See [the Eddystone specification](https://github.com/google/eddystone/tree/master/eddystone-eid) at GitHub.
}
