% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proximitybeacon_objects.R
\name{GetInfoForObservedBeaconsRequest}
\alias{GetInfoForObservedBeaconsRequest}
\title{GetInfoForObservedBeaconsRequest Object}
\usage{
GetInfoForObservedBeaconsRequest(observations = NULL,
  namespacedTypes = NULL)
}
\arguments{
\item{observations}{The beacons that the client has encountered}

\item{namespacedTypes}{Specifies what kind of attachments to include in the response}
}
\value{
GetInfoForObservedBeaconsRequest object
}
\description{
GetInfoForObservedBeaconsRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request for beacon and attachment information about beacons that a mobile client has encountered 'in the wild'.
}
\seealso{
Other GetInfoForObservedBeaconsRequest functions: \code{\link{beaconinfo.getforobserved}}
}
