% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proximitybeacon_objects.R
\name{GetInfoForObservedBeaconsResponse}
\alias{GetInfoForObservedBeaconsResponse}
\title{GetInfoForObservedBeaconsResponse Object}
\usage{
GetInfoForObservedBeaconsResponse(beacons = NULL)
}
\arguments{
\item{beacons}{Public information about beacons}
}
\value{
GetInfoForObservedBeaconsResponse object
}
\description{
GetInfoForObservedBeaconsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Information about the requested beacons, optionally including attachment data.
}
