% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proximitybeacon_objects.R
\name{IndoorLevel}
\alias{IndoorLevel}
\title{IndoorLevel Object}
\usage{
IndoorLevel(name = NULL)
}
\arguments{
\item{name}{The name of this level}
}
\value{
IndoorLevel object
}
\description{
IndoorLevel Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Indoor level, a human-readable string as returned by Google Maps APIs, useful to indicate which floor of a building a beacon is located on.
}
