% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proximitybeacon_objects.R
\name{ListBeaconAttachmentsResponse}
\alias{ListBeaconAttachmentsResponse}
\title{ListBeaconAttachmentsResponse Object}
\usage{
ListBeaconAttachmentsResponse(attachments = NULL)
}
\arguments{
\item{attachments}{The attachments that corresponded to the request params}
}
\value{
ListBeaconAttachmentsResponse object
}
\description{
ListBeaconAttachmentsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response to ListBeaconAttachments that contains the requested attachments.
}
