% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proximitybeacon_objects.R
\name{ListBeaconsResponse}
\alias{ListBeaconsResponse}
\title{ListBeaconsResponse Object}
\usage{
ListBeaconsResponse(beacons = NULL, nextPageToken = NULL,
  totalCount = NULL)
}
\arguments{
\item{beacons}{The beacons that matched the search criteria}

\item{nextPageToken}{An opaque pagination token that the client may provide in their next request to retrieve the next page of results}

\item{totalCount}{Estimate of the total number of beacons matched by the query}
}
\value{
ListBeaconsResponse object
}
\description{
ListBeaconsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response that contains list beacon results and pagination help.
}

