% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proximitybeacon_objects.R
\name{ListDiagnosticsResponse}
\alias{ListDiagnosticsResponse}
\title{ListDiagnosticsResponse Object}
\usage{
ListDiagnosticsResponse(diagnostics = NULL, nextPageToken = NULL)
}
\arguments{
\item{diagnostics}{The diagnostics matching the given request}

\item{nextPageToken}{Token that can be used for pagination}
}
\value{
ListDiagnosticsResponse object
}
\description{
ListDiagnosticsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response that contains the requested diagnostics.
}

