% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proximitybeacon_objects.R
\name{ListNamespacesResponse}
\alias{ListNamespacesResponse}
\title{ListNamespacesResponse Object}
\usage{
ListNamespacesResponse(namespaces = NULL)
}
\arguments{
\item{namespaces}{The attachments that corresponded to the request params}
}
\value{
ListNamespacesResponse object
}
\description{
ListNamespacesResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response to ListNamespacesRequest that contains all the project's namespaces.
}

