% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proximitybeacon_objects.R
\name{Namespace}
\alias{Namespace}
\title{Namespace Object}
\usage{
Namespace(namespaceName = NULL, servingVisibility = NULL)
}
\arguments{
\item{namespaceName}{Resource name of this namespace}

\item{servingVisibility}{Specifies what clients may receive attachments under this namespace via `beaconinfo}
}
\value{
Namespace object
}
\description{
Namespace Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An attachment namespace defines read and write access for all the attachments created under it. Each namespace is globally unique, and owned by one project which is the only project that can create attachments under it.
}
\seealso{
Other Namespace functions: \code{\link{namespaces.update}}
}
