% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proximitybeacon_objects.R
\name{Observation}
\alias{Observation}
\title{Observation Object}
\usage{
Observation(advertisedId = NULL, telemetry = NULL, timestampMs = NULL)
}
\arguments{
\item{advertisedId}{The ID advertised by the beacon the client has encountered}

\item{telemetry}{The array of telemetry bytes received from the beacon}

\item{timestampMs}{Time when the beacon was observed}
}
\value{
Observation object
}
\description{
Observation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents one beacon observed once.
}

