% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proximitybeacon_functions.R
\docType{package}
\name{proximitybeacon_googleAuthR}
\alias{proximitybeacon_googleAuthR}
\alias{proximitybeacon_googleAuthR-package}
\title{Google Proximity Beacon API
Registers, manages, indexes, and searches beacons.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2016-09-03 23:21:46
filename: /Users/mark/dev/R/autoGoogleAPI/googleproximitybeaconv1beta1.auto/R/proximitybeacon_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/userlocation.beacon.registry
}
}

